requirements.txt
numpy==2.0.2
deepface==0.0.91
pillow==10.3.0
filetype==1.2.0
PyYAML==6.0.1
pymongo==4.6.3
SQLAlchemy==2.0.37
python-dotenv==1.0.1
setuptools==69.2.0
opencv-python==4.10.0.84
matplotlib==3.9.0rc2
pandas==2.2.2
Flask==3.1.0
keras==3.8.0
tensorflow==2.18.0
tqdm==4.66.2
requests==2.31.0
gdown==5.1.0
mtcnn==0.1.1